Store dimension size in array metadata
Previously, at least for BASIC, the value stored in the array metadata
was the upper bound of the dimension. This commit changes the value stored
to be the dimension size, which will be consistent with other languages.
The BASIC semantics parser adds one to the specified subscripts in the array
declaration. The generic code generation for array access does not need to
consider the OPTION BASE anymore.