Add new functions (#31)
* Add functions cvd, cvi, mkd$, and mki$

Move functions asc, cdbl, cint, and int to standard library.

* Add standard library functions lbound and ubound

* Store dimension size in array metadata

Previously, at least for BASIC, the value stored in the array metadata
was the upper bound of the dimension. This commit changes the value stored
to be the dimension size, which will be consistent with other languages.
The BASIC semantics parser adds one to the specified subscripts in the array
declaration. The generic code generation for array access does not need to
consider the OPTION BASE anymore.

* Update standard library to 1.4.0