Make more modular (#32)
* Create CompilerFactory and GenericCompiler

Previously, there was one specialized compiler for each language, and class Jcc
was responsible for creating an instance of the one matching the source language.
This commit introduces a CompilerFactory that creates a set of components matching
the source language, and assembles an instance of GenericCompiler from the
components.

Class Jcc is now only responsible for parsing command line arguments and creating
the CompilerFactory. It does not know or care about the source language. Neither
does the GenericCompiler.

This change creates the possibility to add frontends that do not use ANTLR4, and
backends that do not use flat assembler. It also makes it easier to add new source
languages.

* Split Maven project into several Maven modules

There is now one base module, one module for ANTLR4 specific code,
one module for each source language, and one module that puts the
entire compiler collection together.